requirements.txt
discord.py
yt-dlp
PyNaCl
python-dotenv
davey
pydub
